Add base class for rich text components
